services/telegram-bot/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-telegram-bot==20.7
google-cloud-firestore==2.19.0
httpx==0.27.0
python-dotenv==1.0.0
